% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_rds_files.R
\name{clean_rds_files}
\alias{clean_rds_files}
\title{Cleans all RDS Files}
\usage{
clean_rds_files()
}
\value{
\code{logical(1)}
}
\description{
Cleans all RDS Files stored in \code{omlTuneBenchR$rdspath}.
}
