% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_all_files.R
\name{download_all_files}
\alias{download_all_files}
\title{Downloads all available run files}
\usage{
download_all_files(overwrite = TRUE, restrict = TRUE)
}
\arguments{
\item{overwrite}{\code{logical(1)}
Overwrite files or just download missing files?}

\item{restrict}{\code{logical(1)}
Restrict to tasks from paper?}
}
\value{
\code{logical(1)}
}
\description{
Downloads all available run files to the path specified in \code{omlTuneBenchR$rdspath}.
}
