% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_algos.R
\name{get_available_algos}
\alias{get_available_algos}
\title{Returns the algo ids for which there are entries in the database.}
\usage{
get_available_algos()
}
\value{
A vector of algo ids.
}
\description{
Returns the algo ids for which there are entries in the database.
}
