% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_tasks.R
\name{get_available_tasks}
\alias{get_available_tasks}
\title{Returns the task_ids, which are saved in the database and can be queried.}
\usage{
get_available_tasks(restrict = FALSE)
}
\arguments{
\item{restrict}{\link{logical}
Restrict the tasks to those that appear in the following paper:
Automatic Exploration of Machine Learning Experiments on OpenML
Daniel K¨uhn, Philipp Probst, Janek Thomas, Bernd Bischl
https://arxiv.org/pdf/1806.10961.pdf}
}
\value{
A vector of task_ids.
}
\description{
Returns the task_ids, which are saved in the database and can be queried.
}
