% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_inverse_trafo}
\alias{get_inverse_trafo}
\title{This returns the inverse-transformation function specified for this parameter.}
\usage{
get_inverse_trafo(learner_id, param_name)
}
\arguments{
\item{learner_id}{The algorithm name the parameter belongs to.}

\item{param_name}{The parameter name.}
}
\description{
This returns the inverse-transformation function specified for this parameter.
}
