% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nearest_setup.R
\name{get_nearest_setup}
\alias{get_nearest_setup}
\title{This is the core function of this entire API.
It receives a list of algorithm_ids (which will all be treated equally) and a task_id.
The parameter list specifies the parameters to "set".
We will search the OpenML database for the evaluated parameter set, which has the minimal euclidean distance to our given parameters.
We then return the performance of the found point.}
\usage{
get_nearest_setup(learner_id, task_id, parameters)
}
\arguments{
\item{task_id}{A task id, given in numeric form.}

\item{parameters}{A data table of the form data.table(parA = c(valA1, valA2), parB = c(valB1, valB2))}

\item{learner_ids}{A vector of algorithm ids, which will all be treated equally. It is required, that these all have the same name (and parameters).}
}
\value{
A vector of setup ids of the nearest points to the given parameters in the database.
}
\description{
This is the core function of this entire API.
It receives a list of algorithm_ids (which will all be treated equally) and a task_id.
The parameter list specifies the parameters to "set".
We will search the OpenML database for the evaluated parameter set, which has the minimal euclidean distance to our given parameters.
We then return the performance of the found point.
}
