% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paramset_for_omllearner.R
\name{get_paramset_for_omllearner}
\alias{get_paramset_for_omllearner}
\title{Get ParamSet for learner_id}
\usage{
get_paramset_for_omllearner(learner_id)
}
\arguments{
\item{learner_id}{(\code{character(1)})
The learner class (e.g. "classif.glmnet")}
}
\value{
\code{\link{ParamSet}}
}
\description{
Returns a ParamSet Object for a given Learner
}
