% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_runs.R
\name{get_runs}
\alias{get_runs}
\title{Gets table with all runs}
\usage{
get_runs(learner_id, task_id)
}
\arguments{
\item{learner_id}{(\code{character(1)})
The learner (e.g. \code{"classif.rpart"})}

\item{task_id}{(\code{integer(1)})
The task_id (e.g. \code{3})}
}
\value{
\code{data.frame}
}
\description{
Gets table with all runs for a given learner and task.
If the runs are not available locally in \code{omlTuneBenchR$rdspath} they will be downloaded from \code{omlTuneBenchR$remote} first.
}
