% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_task_metadata.R
\name{get_task_metadata}
\alias{get_task_metadata}
\title{Returns the task metadata table.}
\usage{
get_task_metadata()
}
\value{
\code{\[data.table\]} Task metadata.
}
\description{
Returns the task metadata table.
}
