% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_task_metadata.R, R/helpers.R
\name{get_task_metadata}
\alias{get_task_metadata}
\title{Returns the task metadata table.}
\usage{
get_task_metadata(task_id)

get_task_metadata(task_id)
}
\arguments{
\item{task_id}{\code{numeric(n)} The task of interest}
}
\value{
\code{\[data.table\]} Task metadata.

nrow and ncol of the dataset.
}
\description{
Returns the task metadata table.

Retrieves metadata about the task from openml.org
}
