% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_task_metadata}
\alias{get_task_metadata}
\title{Retrieves metadata about the task from openml.org}
\usage{
get_task_metadata(task_id)
}
\arguments{
\item{task_id}{\code{numeric(n)} The task of interest}
}
\value{
nrow and ncol of the dataset.
}
\description{
Retrieves metadata about the task from openml.org
}
