% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_omlbenchfunction.R
\name{make_omlbenchfunction}
\alias{make_omlbenchfunction}
\title{Creates a function that will wrap the API}
\usage{
make_omlbenchfunction(learner_id, task_id, include.extras = FALSE,
  objective = "accuracy")
}
\arguments{
\item{learner_id}{(\code{character(1)})
Defines the learner. (e.g. \code{"classif.rpart"})}

\item{task_id}{(\code{integer(1)})
Defines the task_id the learner is optimized on.}

\item{include.extras}{(\code{logical(1)})
Should the extras be attached as an attribute?}

\item{objective}{(\code{character(1)})
Which value should be optimized? Possible choices are \dQuote{auc}, \dQuote{accuracy} (default) and \dQuote{rmse}.}
}
\value{
\code{function}
}
\description{
Wraps the API Call in a function that returns the performance of a defined learner and \code{task_id} for a given \code{x} using the OmlBench API.
}
